По возможности убран дублирующийся код
